+++ src/main.js
@@ -1,4 +1,5 @@
 import { createApp } from 'vue'
+import { i18n } from '@/i18n'
 import { router } from '@/router'
 import { store } from '@/stores'
 import App from '@/App.vue'
@@ -7,6 +8,7 @@ import '@/styles/main.scss'
 
 const Finances = createApp(App)
 
+Finances.use(i18n)
 Finances.use(router)
 Finances.use(store)