+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.2.0
+
+* Add internationalization
+  * Spanish locale (default)
+  * English locale
+
 ## 1.1.2
 
 * Register an expense whenever a transfer between accounts is made